[Task] Make sure tasks belong to user before editing / deleting

+++ app/Http/Controllers/TaskController.php
@@ -15,18 +15,22 @@ class TaskController extends Controller
             'due_by' => 'required|date_format:Y-m-d',
         ]);
 
-        Task::create($validated);
+        auth()->user()->tasks()->create($validated);
 
         return redirect()->back()->with('success', 'Task created successfully.');
     }
 
     public function update(Request $request, Task $task)
     {
+        if($task->user_id !== auth()->user()->id) {
+            return redirect()->back()->with('error', 'This task does not belong to you!.');
+        }
+
         $validated  = $request->validateWithBag('taskUpdate', [
-            'name' => 'required|string|max:180',
-            'description' => 'required|string|max:500',
-            'status' => 'required|in:pending,complete,overdue',
-            'due_by' => 'required|date_format:d/m/Y',
+            'name' => 'sometimes|string|max:180',
+            'description' => 'sometimes|string|max:500',
+            'status' => 'sometimes|in:pending,complete,overdue',
+            'due_by' => 'sometimes|date_format:d/m/Y',
         ]);
 
         $task->update($validated);
@@ -36,6 +40,10 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        if($task->user_id !== auth()->user()->id) {
+            return redirect()->back()->with('error', 'This task does not belong to you!.');
+        }
+
         $task->delete();
         return redirect()->back()>with('success', 'Task deleted successfully.');
     }
